revert(settings): restore the settings sidebar, keep only the breadcrumb change [ENG-1705]
Reverts the sidebar work from this branch (e51f5f6 and acd9fe2) and keeps
059db1a. Scope is now solely the top bar: on the workspace-agnostic settings
routes the workspace breadcrumb is not rendered, and the sidebar is byte-for-byte
what it was before this branch.

Restored to their pre-branch state:

- SettingsSidebarContent, including its hideWorkspaceSection prop and the
  Workspace section on organization and account settings
- MainNavigation and settings-navigation, which render it
- settings-shell's workspaceId/workspaceName wiring
- navigation-data's comments

Also removes modules/settings/components/sidebar/, which only existed to split
that component in two.

The spec is rewritten and renamed to match the narrower scope: it now asserts
the sidebar Workspace section is *present* on every settings route -- the guard
that the breadcrumb change does not leak into the sidebar -- while the workspace
breadcrumb is absent on the workspace-agnostic routes and present on the
workspace-scoped ones.
