+++ apps/web/modules/settings/components/settings-navigation.tsx
@@ -5,14 +5,19 @@ import { useCallback, useTransition } from "react";
 import { useTranslation } from "react-i18next";
 import type { TOrganizationRole } from "@formbricks/types/memberships";
 import type { TUser } from "@formbricks/types/user";
-import { getOrganizationsForSwitcherAction } from "@/app/(app)/workspaces/[workspaceId]/actions";
-import { OrganizationSettingsSidebar } from "@/modules/settings/components/sidebar/organization-settings-sidebar";
+import {
+  getOrganizationsForSwitcherAction,
+  getWorkspacesForSwitcherAction,
+} from "@/app/(app)/workspaces/[workspaceId]/actions";
+import { SettingsSidebarContent } from "@/app/(app)/workspaces/[workspaceId]/components/SettingsSidebarContent";
 import { UserDropdown } from "@/modules/settings/components/user-dropdown";
 import { useSwitcherData } from "@/modules/settings/hooks/use-switcher-data";
 import { GoBackButton } from "@/modules/ui/components/go-back-button";
 
 interface SettingsNavigationProps {
   user: TUser;
+  workspaceId: string;
+  workspaceName: string;
   organizationId: string;
   organizationName: string;
   membershipRole?: TOrganizationRole;
@@ -23,13 +28,13 @@ interface SettingsNavigationProps {
   backUrl: string;
 }
 
-// The settings-mode sidebar chrome for the workspace-agnostic settings routes
-// (/organizations/[organizationId]/settings and /account/settings). It renders
-// OrganizationSettingsSidebar, which shares its Organization and Account sections with the
-// in-workspace sidebar (MainNavigation -> WorkspaceSettingsSidebar) but has no Workspace section:
-// these routes carry no workspaceId, and the top bar breadcrumb is where workspaces get switched.
+// The settings-mode sidebar chrome, extracted so it can render both inside the workspace layout
+// (MainNavigation) and on the standalone /organizations/[id]/settings and /account/settings routes —
+// one component, identical UI, regardless of which scoped route is active.
 export const SettingsNavigation = ({
   user,
+  workspaceId,
+  workspaceName,
   organizationId,
   organizationName,
   membershipRole,
@@ -42,9 +47,22 @@ export const SettingsNavigation = ({
   const { t } = useTranslation();
   const [, startTransition] = useTransition();
 
+  const workspaceSwitcher = useSwitcherData(
+    () => getWorkspacesForSwitcherAction({ organizationId }),
+    t("common.failed_to_load_workspaces")
+  );
   const organizationSwitcher = useSwitcherData(
     () => getOrganizationsForSwitcherAction({ organizationId }),
     t("common.failed_to_load_organizations")
+  );
+
+  const handleWorkspaceChange = useCallback(
+    (id: string) => {
+      startTransition(() => {
+        router.push(`/workspaces/${id}/settings/workspace/general`);
+      });
+    },
+    [router]
   );
 
   const handleOrganizationChange = useCallback(
@@ -64,13 +82,24 @@ export const SettingsNavigation = ({
         <div className="mb-2 px-3">
           <GoBackButton url={backUrl} />
         </div>
-        <OrganizationSettingsSidebar
+        <SettingsSidebarContent
+          workspaceId={workspaceId}
+          workspaceName={workspaceName}
           organizationId={organizationId}
           organizationName={organizationName}
           membershipRole={membershipRole}
           isFormbricksCloud={isFormbricksCloud}
           isCollapsed={false}
           isTextVisible={false}
+          hideWorkspaceSection={!workspaceId}
+          workspaces={workspaceSwitcher.items}
+          isLoadingWorkspaces={workspaceSwitcher.isLoading}
+          onWorkspaceChange={handleWorkspaceChange}
+          onWorkspaceDropdownOpen={() =>
+            workspaceSwitcher.error ? workspaceSwitcher.retry() : workspaceSwitcher.load()
+          }
+          errorWorkspaces={workspaceSwitcher.error}
+          onWorkspaceRetry={workspaceSwitcher.retry}
           organizations={organizationSwitcher.items}
           isLoadingOrganizations={organizationSwitcher.isLoading}
           onOrganizationChange={handleOrganizationChange}
